Переделал задание из module6-task2: перенёс конструктор Review в модуль review, переписал вызов конструктора Review в модуле reviews, удалил лишний модуль review-constructor
