+++ src/js/review.js
@@ -6,7 +6,7 @@
  * @param {Array} reviews
  * @returns {Node} reviewItem
  */
-module.exports = function(reviews) {
+var getReviewItem = function(reviews) {
   var template = document.getElementById('review-template'),
     templateContainer = 'content' in template ? template.content : template,
     reviewItem = templateContainer.cloneNode(true),
@@ -29,3 +29,36 @@ module.exports = function(reviews) {
   return reviewItem;
 };
 
+var Review = function(data) {
+  this.data = data;
+  this.element = getReviewItem(this.data);
+
+  var self = this;
+
+  /**
+   * @const
+   * @type {string}
+   */
+  var CLASS_ACTIVE = 'review-quiz-answer-active';
+
+  this.answers = this.element.querySelectorAll('.review-quiz-answer');
+
+  Array.prototype.forEach.call(this.answers, function(answer) {
+    answer.addEventListener('click', function(evt) {
+      for (var i = 0; i < self.answers.length; i++) {
+        self.answers[i].classList.remove(CLASS_ACTIVE);
+      }
+      evt.target.classList.add(CLASS_ACTIVE);
+    });
+  });
+};
+
+Review.prototype = {
+  remove: function() {
+    Array.prototype.forEach.call(this.answers, function(answer) {
+      answer.removeEventListener('click');
+    });
+  }
+};
+
+module.exports = Review;
